requirements.txt
ultralytics
huggingface_hub
opencv-python
numpy
pdf2image
Pillow
scikit-image
matplotlib
